testing/requirements.txt
torch
transformers>=4.49.0
accelerate>=0.26.0
bitsandbytes>=0.42.0
olmocr>=0.4.0
pypdf>=4.0.0
Pillow>=10.0.0
pyyaml>=6.0.0
qwen-vl-utils
# Tokenization and Model Support
sentencepiece
einops
# Note: Ensure poppler-utils is installed on the system for PDF rendering
